Patrón Factory Method Aplicado

+++ app/src/main/java/com/dannyang27/sportpoints/activities/PruebasDanny/CustomPartAdapter.java
@@ -34,7 +34,7 @@ public class CustomPartAdapter extends ArrayAdapter<Participante> {
 
         Participante p = getItem(position);
         nombre.setText(p.getNombre());
-        tlf.setText(p.getTelefono());
+        tlf.setText((int) p.getTelefono());
 
 
         return convertView;
